requirements.txt
appdirs==1.4.3
asgiref==3.2.5
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.10
virtualenvwrapper-win==1.2.6
